Health decreases when enemy touches player
