--- Assets/SchoolRPG/Source/SchoolRPG/health/Death.cs
@@ -1,20 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-    public class Death : MonoBehaviour
-    {
-        // Start is called before the first frame update
-        void Start()
-        {
-
-        }
-        public void die()
-        {
-        }
-        // Update is called once per frame
-        void Update()
-        {
-
-        }
-    }
