--- Assets/SchoolRPG/Source/SchoolRPG/health/Death.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 76e220afb9c63464485c2141bca9bda0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
