--- Assets/SchoolRPG/Source/SchoolRPG/health/Temporary_Change_Color.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7a3dd596775587041b83851361d11f7f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
