--- Assets/SchoolRPG/Source/SchoolRPG/health/health.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b8e1a73978ed0748adc245cbd35c0be
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
